[#163] CR PR fixes

+++ cpp/src/robotick/workloads/telemetry/MqttClientWorkload.cpp
@@ -20,12 +20,12 @@ namespace robotick
 	struct MqttClientConfig
 	{
 		FixedString64 broker_url = "mqtt://localhost";
-		int broker_mqtt_port = 1883;
+		uint16_t broker_mqtt_port = 1883;
 		FixedString64 root_topic_namespace = "robotick";
 	};
 	ROBOTICK_REGISTER_STRUCT_BEGIN(MqttClientConfig)
 	ROBOTICK_STRUCT_FIELD(MqttClientConfig, FixedString64, broker_url)
-	ROBOTICK_STRUCT_FIELD(MqttClientConfig, int, broker_mqtt_port)
+	ROBOTICK_STRUCT_FIELD(MqttClientConfig, uint16_t, broker_mqtt_port)
 	ROBOTICK_STRUCT_FIELD(MqttClientConfig, FixedString64, root_topic_namespace)
 	ROBOTICK_REGISTER_STRUCT_END(MqttClientConfig)
